add fallback targets if the main target is not found
